Rewriting some output functions and disabling others
